修改了arctan函数，优化了界面
