--- arcsin.h
@@ -1,12 +1,0 @@
-#ifndef ARCSIN_H
-#define ARCSIN_H
-
-
-class ArcSin
-{
-public:
-    ArcSin();
-    double MyArcsin(double dous,bool isAngle);
-};
-
-#endif // ARCSIN_H
