+++ include/arcsin.h
@@ -1,0 +1,12 @@
+#ifndef ARCSIN_H
+#define ARCSIN_H
+#include <qstring.h>
+
+class ArcSin
+{
+public:
+    ArcSin();
+    QString ArcsinTest(QString sIn);
+};
+
+#endif // ARCSIN_H
