+++ include/cos.h
@@ -1,0 +1,13 @@
+#ifndef COS_H
+#define COS_H
+#include <qstring.h>
+
+
+class Cos
+{
+public:
+    Cos();
+    QString CinTest(QString sIn);
+};
+
+#endif // COS_H
