+++ include/sin.h
@@ -1,0 +1,12 @@
+#ifndef SIN_H
+#define SIN_H
+#include <qstring.h>
+
+class Sin
+{
+public:
+    Sin();
+    QString SinTest(QString sIn);
+};
+
+#endif // SIN_H
