--- main.cpp
@@ -1,11 +1,0 @@
-#include "calyh.h"
-#include <QApplication>
-
-int main(int argc, char *argv[])
-{
-    QApplication a(argc, argv);
-    CalYh w;
-    w.show();
-
-    return a.exec();
-}
