--- sin.h
@@ -1,12 +1,0 @@
-#ifndef SIN_H
-#define SIN_H
-
-
-class Sin
-{
-public:
-    Sin();
-    double sin(double num);
-};
-
-#endif // SIN_H
